Refactor class T from core/worker and plugins/camera_control to utils
